pyproject.toml
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "Citadel"
version = "0.1.0"
description = "Discord bot for your next study jam"
authors = ["Hunter Wittenborn <hunter@hunterwittenborn.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "3.12.*"
rich = "^13.7.1"
mypy = "^1.11.0"
python-dotenv = "^1.0.1"
discord-py = "^2.4.0"
typer = "^0.12.3"
openai = "^1.37.0"
jinja2 = "^3.1.4"

[tool.poetry.dev-dependencies]
ruff = "~0.5.0"
pre-commit = "~3.7.1"

[tool.poetry.scripts]
citadel = "citadel.main:entrypoint"

[tool.ruff]
line-length = 119
target-version = "py312"
fix = true
src = ["src"]

[tool.ruff.lint]
select = ["ALL"]
ignore = [
    "D100",
    "D104",
    "D105",
    "D106",
    "D107",
    "D203",
    "D213",
    "D415",
    "D301",
    "A",
    "T20",
    "TD002",
    "TD003",
    "FIX",
    "ANN101",
    "ANN102",
]
extend-select = ["I"]
